http_image_filter_module: Add HEIC and AVIF support
Libgd added support for HEIC (HEVC encoded images in HEIF containers)
and AVIF (AV1 encoded images in HEIF containers).
Support for this formats is added to the image_filter_module similarly
as with the WebP format.
HEIC is used as the name for the fileformat containing HEVC inside HEIF
containers to avoid confusion between HEIF, HEIC and AVIF.
These new formats improve compression efficiency over older formats such as
JPEG.
At the moment browser support is in development and it is unclear if and which
format will dominate usage on the web.
Therefore, support for both formats is added, which is straightforward when
using libgd.

The following new configuration commands are added:

* image_filter_heic_quality: 0-100 or 200 for lossless
  Larger values result in better quality and larger file size.
* image_filter_avif_quality: 0-100
  Larger values result in better quality and larger file size.
* image_filter_avif_speed: 0-10
  Larger values result in faster encoding and lower compression ratio.

If the options are not set, -1 is passed to libgd to use the default value.

The autotools tests for WebP, HEIC and AVIF support in libgd have been fixed.
Libgd without support for these formats includes the API functions, but
calling them always results in an error.
Therefore, the test compilation always succeeds even if the format is not
supported.
The new test must be executed and calls the libgd check for filetype support.

http_image_filter_module: Add output format option

This change makes it possible to specify the output format of the generated
images.
The option can be used in combination with other filters or standalone to
simply convert the image.
As such one can generate images in more efficiently compressed
image formats without manually convert and store all format files.
This allows to serve multiple image formats and specify the alternatives
in an HTML picture tag. The browser can then select an appropriate format.
A caching server can be used to avoid reconverting the same image multiple
times.

The following filter command allows to configure the output format:

* image_filter convert TYPE
  If no other filter is set, then the unchanged image will be converted to the
  output format.
  Possible values for TYPE: input, jpeg, gif, png, webp, heic, avif.
  The default value is: input.
  For the value `input` the format of the input file is used, which is
  the behavior before the change.

http_image_filter_module: Fix conversion in case the image is not resized